% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressed_genes.R
\name{add_expressed_genes}
\alias{add_expressed_genes}
\title{Adding the nearest expressed genes}
\usage{
add_expressed_genes(
  input_file = NULL,
  tissue,
  gtf_path,
  species = "Homo_sapiens",
  annot_ers
)
}
\value{
Granges with annotated ERs and details of their nearest expressed
genes
}
\description{
Adding the nearest gene and nearest expressed gene to the mcols of the
annotated ERs
}
\examples{
\dontshow{
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
) # .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache

bw_path <- ODER:::.file_cache(url[1])
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
gtf_path <- ODER:::.file_cache(gtf_url)
}

opt_ers <- ODER(
    bw_paths = bw_path, auc_raw = auc_example,
    auc_target = 40e6 * 100, chrs = c("chr21", "chr22"),
    genome = "hg38", mccs = c(5, 10), mrgs = c(10, 20),
    gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE,
    exons_no_overlap = NULL, bw_chr = "chr"
)

junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)
annot_ers <- annotatERs(
    opt_ers = opt_ers[["opt_ers"]], junc_data = junctions,
    gtf_path = gtf_path, chrs_to_keep = c("21", "22"), ensembl = TRUE
)

annot_ers <- add_expressed_genes(
    tissue = "lung", gtf_path = gtf_path,
    annot_ers = annot_ers
)
}
