% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auc_example}
\alias{auc_example}
\title{An example AUC value}
\format{
A numeric value
}
\source{
See example.R in data-raw
}
\usage{
auc_example
}
\description{
An AUC(Area Under Coverage) value for a user to try out the package and to
pass in for tests. From the GTEX data set and project SRP012682, the actual
value is 11872688252.
}
\keyword{datasets}
