% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ERs.R
\name{calc_mean_cov}
\alias{calc_mean_cov}
\title{Collapse bigwig files into a mean coverage}
\usage{
calc_mean_cov(bw_paths, aucs, target_auc, chr_to_filter)
}
\arguments{
\item{bw_paths}{paths to bigwig files.}

\item{aucs}{vector containing aucs matching the order of bigwig paths.}

\item{target_auc}{total auc to normalise all samples to. E.g. 40e6 * 100
would be the estimated total AUC for sample sequenced to 40 million reads
of 100bp in length.}

\item{chr_to_filter}{chrs to obtain mean coverage for.}
}
\value{
list of Rle objects. Each containing the mean coverage for an entire
  chromosome.
}
\description{
\code{calc_mean_cov} takes as input a set of bws. Then uses the total AUC to
normalise coverage per sample and finally collapses the coverage for each
sample into a mean coverage.
}
