% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverage_example}
\alias{coverage_example}
\title{An example coverage}
\format{
A list of length 2 containing 2 Rles for chromosomes 21 and 22 respectively
}
\source{
See example.R in data-raw
}
\usage{
coverage_example
}
\description{
An example coverage generated for a user to try out the package and
to pass in for tests
}
\keyword{datasets}
