% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-ERs.R
\name{.delta}
\alias{.delta}
\title{Default delta functions}
\usage{
.delta(query, subject)
}
\arguments{
\item{query}{Set of ERs}

\item{subject}{Optimum exons}
}
\value{
summarised delta scores
}
\description{
Default delta functions
}
\keyword{internal}
