% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ers_example}
\alias{ers_example}
\title{An example set of Expressed Regions}
\format{
A list containing two lists (for each mcc) each with a set of genomic ranges
for the different combinations of mcc and mrg
}
\source{
See example.R in data-raw
}
\usage{
ers_example
}
\description{
An example set of Expressed Regions generated for a user to try out the
package and to pass in for tests. Generated using coverage_example and MCCs
of 5 and 10 and MRGs of 10 and 20.
}
\keyword{datasets}
