% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{file_cache}
\alias{file_cache}
\title{Cache a file if it is not found locally}
\usage{
file_cache(file_path)
}
\arguments{
\item{file_path}{a path to file of interest.}
}
\value{
file_path of cached file or unchanged file_path if found locally.
}
\description{
\code{file_cache} will use \link[BiocFileCache:BiocFileCache-class]{BiocFileCache}
will cache the file for faster repeated retrival, if it is not found locally
(i.e. a URL).
}
\examples{
if (!exists("rec_url")) {
    rec_url <- recount::download_study(
        project = "SRP012682",
        type = "samples",
        download = FALSE
    )
}
eg_bwfile <- file_cache(rec_url[1])
eg_bwfile
}
