% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ERs.R
\name{gen_ERs}
\alias{gen_ERs}
\title{Define sets of ERs}
\usage{
gen_ERs(chrs_mean_cov, MCCs, MRGs)
}
\arguments{
\item{chrs_mean_cov}{output of \code{\link{cal_mean_cov}}.}

\item{MCCs}{mean coverage cut-offs to apply.}

\item{MRGs}{max region gaps to apply.}
}
\value{
list containing sets of ERs, each generated using a particular
  combination of MCC and MRG.
}
\description{
\code{gen_ERs} defines ERs across a inputted range of mean coverage cut-offs
(MCCs) and max region gaps (MRGs).
}
