% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatER.R
\name{generate_genomic_state}
\alias{generate_genomic_state}
\title{Generating a genomic state object from Txdb or gtf}
\usage{
generate_genomic_state(
  gtf = NULL,
  txdb = NULL,
  chrs_to_keep = c(1:22, "X", "Y", "MT"),
  ensembl = TRUE
)
}
\arguments{
\item{gtf}{gtf file, will be converted into a txdb and then genomic state, if
no gtf is entered a txdb must be}

\item{txdb}{txdb object, if one is not entered a gtf file needs to be}

\item{chrs_to_keep}{chromosomes to keep in genomic state (in NCBI format i.e.
1,2,3...22,X,Y,MT)}

\item{ensembl}{logical variable to say whether the gtf file entered is from
ensembl, the default is true}
}
\value{
a genomic state
}
\description{
\code{generate_genomic_state} takes txdb object or a gtf file and makes a
genomic state
}
\examples{
\dontshow{
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
# .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache
gtf_path <- ODER:::.file_cache(gtf_url)
}
genom_state <- generate_genomic_state(
    gtf = gtf_path,
    chrs_to_keep = c("1", "2", "X"), ensembl = TRUE
)
}
