% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-ERs.R
\name{get_coverage}
\alias{get_coverage}
\title{Generating the mean coverage of the expressed regions}
\usage{
get_coverage(bw_paths, auc_raw, auc_target, chrs = "", genome = "hg38")
}
\arguments{
\item{bw_paths}{paths to bigwig files with the RNA-seq data that you want the coverage of.}

\item{auc_raw}{vector containing aucs matching the order of bigwig paths.}

\item{auc_target}{total auc to normalise all samples to. E.g. 40e6 * 100
would be the estimated total auc for sample sequenced to 40 million reads
of 100bp in length.}

\item{chrs}{chromosomes to obtain mean coverage for, default is "" giving every chromosome}

\item{genome}{the UCSC genome you want to use, the default is hg38}
}
\value{
a list of Rles per chromosome passed in
}
\description{
\code{get_coverage} returns the coverage of the BigWig data passed in
}
\examples{
eg_coverage <- get_coverage(
    # 	bw_paths = bw_path,
    auc_raw = gtex_metadata[["auc"]][1],
    auc_target = 40e6 * 100,
    chrs = test_chrs
)
}
