% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-ERs.R
\name{get_ers_delta}
\alias{get_ers_delta}
\title{Calculates delta for sets of ERs}
\usage{
get_ers_delta(ers, opt_exons, delta_fun = .delta)
}
\arguments{
\item{ers}{Sets of ERs across various MCCs/MRGs - output of
\code{\link{get_ers}}.}

\item{opt_exons}{GRanges object that contains the regions that ideally, you
want to the ER definitions to match - output of \code{\link{get_exons}}.}

\item{delta_fun}{Function that calculates the delta between ERs and
\code{opt_exons}. Takes as input a set of ERs from \code{ers} and
\code{opt_exons}. Then outputs a tibble/dataframe containing the summarised
delta scores for that set of one set of ERs.}
}
\value{
tibble/dataframe containing summarised delta values. One row per set
of ERs.
}
\description{
Calculates delta for sets of ERs
}
\examples{
gtf_url <- "http://ftp.ensembl.org/pub/release-103/gtf/homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
gtf_path <- ODER:::.file_cache(gtf_url)

eg_opt_exons <- get_exons(gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE)

eg_ers_delta <- get_ers_delta(ers = ers, opt_exons = eg_opt_exons, delta_fun = .delta)

print(eg_ers_delta)
}
