% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressed_genes.R
\name{get_expressed_genes}
\alias{get_expressed_genes}
\title{Get the expressed genes}
\usage{
get_expressed_genes(gtf_path, species = "Homo_sapiens", tissue_df)
}
\arguments{
\item{gtf_path}{gtf file path}

\item{species}{character string containing the species to filter for,
Homo sapiens is the default}

\item{tissue_df}{dataframe containing the expressed genes for a particular
tissue}
}
\value{
GRanges with the expressed genes for a specific tissue
}
\description{
Takes in a gtf file and a dataframe with expressed genes from a specific
tissue. Filters the gtf file for genes and keeps those that are present in the
tissue dataframe passed in.
}
\examples{
\dontshow{
if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    # .file_cache is an internal function to download a bigwig file from a link
    # if the file has been downloaded recently, it will be retrieved from a cache
    gtf_path <- .file_cache(gtf_url)
}
}
lung_tissue <- get_tissue(tissue = "lung")

lung_exp_genes <- get_expressed_genes(
    gtf_path = gtf_path,
    tissue_df = lung_tissue
)

lung_exp_genes
}
