% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-ERs.R
\name{get_opt_ers}
\alias{get_opt_ers}
\title{Obtains optimised set of ERs}
\usage{
get_opt_ers(ers, ers_delta)
}
\arguments{
\item{ers}{Sets of ERs across various MCCs/MRGs - output of
\code{\link{get_ers}}.}

\item{ers_delta}{tibble/dataframe containing summarised delta values. One row per set
of ERs.}
}
\value{
list containing optimised ERs, optimal pair of MCC/MRGs and
\code{delta_df}
}
\description{
Obtains optimised set of ERs
}
\examples{

gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
gtf_path <- ODER:::.file_cache(gtf_url)
exons_no_overlap <- get_exons(
    gtf = gtf_path,
    ucsc_chr = TRUE,
    ignore.strand = TRUE
)

ers_delta <- get_ers_delta(
    ers = ers_example,
    opt_exons = exons_no_overlap,
    delta_fun = .delta
)

opt_ers <- get_opt_ers(
    ers = ers_example,
    ers_delta = ers_delta
)
}
