% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ers.R
\name{get_strand_ers}
\alias{get_strand_ers}
\title{Define sets of ERs for stranded Bigwigs}
\usage{
get_strand_ers(
  bw_pos,
  bw_neg,
  auc_raw_pos,
  auc_raw_neg,
  auc_target,
  chrs,
  mccs,
  mrgs,
  bw_chr = "chr"
)
}
\arguments{
\item{bw_pos}{positive strand bigwig file}

\item{bw_neg}{negative strand bigwig file}

\item{auc_raw_pos}{vector containing AUCs(Area Under Coverage) matching the
order of the positive bigwig paths.}

\item{auc_raw_neg}{vector containing AUCs(Area Under Coverage) matching the
order of the negative bigwig paths.}

\item{auc_target}{total AUC to normalise all samples to. E.g. 40e6 * 100
would be the estimated total auc for sample sequenced to 40 million reads
of 100bp in length.}

\item{chrs}{chromosomes to obtain mean coverage for, default is "" giving
every chromosome. Can take UCSC format(chrs = "chr1")
or just the chromosome i.e. chrs = c(1,X)}

\item{mccs}{mean coverage cut-offs to apply.}

\item{mrgs}{max region gaps to apply.}

\item{bw_chr}{specifies whether the bigwig files has the chromosomes labelled
with a "chr" preceding the chromosome i.e. "chr1" vs "1". Can be either
"chr" or "nochr" with "chr" being the default.}
}
\value{
list containing sets of stranded ERs, each generated using a
particular combination of MCC and MRG.
}
\description{
\code{get_strand_ers} defines ERs across an inputted range of mean coverage
cut-offs (MCCs) and max region gaps (MRGs) from the coverage.
}
\examples{
library("magrittr")
if (!exists("gtex_metadata")) {
    gtex_metadata <- recount::all_metadata("gtex")
    gtex_metadata <- gtex_metadata \%>\%
        as.data.frame() \%>\%
        dplyr::filter(project == "SRP012682")
}
if (!exists("rec_url")) {
    rec_url <- recount::download_study(
        project = "SRP012682",
        type = "samples",
        download = FALSE
    )
}
# file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache
bw_plus <- file_cache(rec_url[58])
bw_minus <- file_cache(rec_url[84])

# As of rtracklayer 1.25.16, BigWig is not supported on Windows.
if (!xfun::is_windows()) {
    stranded_ers <- get_strand_ers(
        bw_pos = bw_plus, bw_neg = bw_minus,
        auc_raw_pos = gtex_metadata[["auc"]][58],
        auc_raw_neg = gtex_metadata[["auc"]][84], auc_target = 40e6 * 100,
        chrs = "chr21", mccs = c(5, 10), mrgs = c(10, 20)
    )
    message(stranded_ers)
}
}
