% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressed_genes.R
\name{get_tissue}
\alias{get_tissue}
\title{Get gene data for a tissue}
\usage{
get_tissue(input_file = NULL, tissue)
}
\arguments{
\item{input_file}{GTEX median expression file, if left as NULL
\code{\link{get_tissue}} will use the default file}

\item{tissue}{Tissue to filter for. See tissue_options for options}
}
\value{
Dataframe containing expressed genes
}
\description{
Generate a dataframe for the tissue entered in containing a list of expressed
genes that appear above a threshold of RPKM>0.1.
(RPKM - Reads Per Kilobase of transcript, per Million mapped reads (RPKM) is
a normalized unit of transcript expression.)
}
\details{
Can take a GTEX median expression file as an input or use the default (v6)
}
\examples{

lung_tissue <- get_tissue(tissue = "lung")

lung_tissue
}
