% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_SRP012682_SRX222703_lung_coverage_1}
\alias{gtex_SRP012682_SRX222703_lung_coverage_1}
\title{An example object containing coverage}
\format{
A list of length 2 containing 2 Rles for chromosomes 21 and
22 respectively
}
\source{
See example.R in data-raw
}
\usage{
gtex_SRP012682_SRX222703_lung_coverage_1
}
\description{
Coverage generated for a user to try out the package and
to pass in for tests. Coverage of chromosomes 21 and 22 from the project
SRP012682.
}
\keyword{datasets}
