% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_SRP012682_SRX222703_lung_erdelta_1}
\alias{gtex_SRP012682_SRX222703_lung_erdelta_1}
\title{An example set of ER deltas}
\format{
A tibble/dataframe with the sums, means, medians, n_eq_0 and
propor_eq_0 for each combination of mccs (5 & 10) and mrgs (10 & 20)
}
\source{
See example.R in data-raw
}
\usage{
data(gtex_SRP012682_SRX222703_lung_erdelta_1)
}
\description{
This set of deltas was calculated using gtex_lung_ers_1 and exons from
ensembl.
}
\keyword{datasets}
