% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_SRP012682_SRX222703_lung_ers_1}
\alias{gtex_SRP012682_SRX222703_lung_ers_1}
\title{An example set of Expressed Regions}
\format{
A list containing two lists (for each mcc) each with a set of genomic
ranges for the different combinations of mcc and mrg
}
\source{
See example.R in data-raw
}
\usage{
gtex_SRP012682_SRX222703_lung_ers_1
}
\description{
An example set of Expressed Regions generated for a user to try out the
package and to pass in for tests. Generated using
gtex_SRP012682_SRX222703_lung_coverage_1 and MCCs of 5 and 10 and MRGs of
10 and 20.
}
\keyword{datasets}
