% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_junc_21_22}
\alias{lung_junc_21_22}
\title{Junction data of chromosomes 21 and 22 from a lung tissue sample}
\format{
A dataframe with the junction ID, chromosome, start and ends, strand,
number of samples, acceptor and donor
}
\source{
GTEx
}
\usage{
data(lung_junc_21_22)
}
\description{
These junctions were sampled from a local junction file.
}
\keyword{datasets}
