% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ERs.R
\name{plot_ers}
\alias{plot_ers}
\title{Plot Expressed regions}
\usage{
plot_ers(ers_delta, opt_mcc_mrg)
}
\arguments{
\item{ers_delta}{tibble/dataframe containing summarised delta values. One row
per set of ERs.}

\item{opt_mcc_mrg}{vector containing the optimum mcc and mrg, in that order}
}
\value{
Plot of MCC against median delta and number of ERS with a delta of 0
}
\description{
Plots the median deltas and the number of ERs with a delta of 0 against the
MCCs on two separate graphs with a line for each of the various MRGs.
}
\examples{
# gtex_SRP012682_SRX222703_lung_erdelta_1 is from the package data folder
eg_plots <- plot_ers(ers_delta = gtex_SRP012682_SRX222703_lung_erdelta_1, opt_mcc_mrg = c(
    "mcc_10",
    "mrg_20"
))

eg_plots
}
