% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudogene}
\alias{pseudogene}
\title{Different transcript biotypes that count as pseudogene}
\format{
A character vector with all of the different pseudogene categories
\code{\link{get_exons}} function.
}
\source{
See exon_biotypes.R in data-raw
}
\usage{
pseudogene
}
\description{
These are the various transcript biotypes typically found in the transcript
biotype column of a gtf file.
}
\keyword{datasets}
