% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine-ERs.R
\name{refine_ERs}
\alias{refine_ERs}
\title{Refines the ERs start and end points}
\usage{
refine_ERs(annot_ers)
}
\arguments{
\item{annot_ers}{ERs that have been annotated (result of annotatER)}
}
\value{
Genomic ranges with refined base pair starts and ends and a logical
vector listing changes
}
\description{
Uses the junctions added by \code{\link{annotatERs}} to modify the starts and
ends of the expressed regions. When a junction intersects an expressed region
depending on whether it is the start or end or both the regions corresponding
starts and ends will be modified.
}
\details{
As junctions mark intron boundaries, the expressed region will be changed to
either being one less or more than the junction end.
}
\examples{
\dontshow{

if (!exists("rec_url")) {
    rec_url <- recount::download_study(
        project = "SRP012682",
        type = "samples",
        download = FALSE
    ) # .file_cache is an internal function to download a bigwig file from a link
    # if the file has been downloaded recently, it will be retrieved from a cache
}
bw_path <- .file_cache(rec_url[1])
if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    gtf_path <- .file_cache(gtf_url)
}
}
if (!exists("opt_ers1")) {
    opt_ers1 <- ODER(
        bw_paths = bw_path, auc_raw = auc_example,
        auc_target = 40e6 * 100, chrs = c("chr21"),
        genome = "hg38", mccs = c(5, 10), mrgs = c(10, 20),
        gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE,
        exons_no_overlap = NULL, bw_chr = "chr"
    )
}
if (!exists("genom_state")) {
    genom_state <- generate_genomic_state(
        gtf = gtf_path,
        chrs_to_keep = c("21"), ensembl = TRUE
    )
}

junctions <- lung_junc_21_22
if (!exists("annot_ers1")) {
    annot_ers1 <- annotatERs(
        opt_ers = head(opt_ers1[["opt_ers"]], 100), junc_data = junctions,
        gtf_path = gtf_path, chrs_to_keep = c("21"), ensembl = TRUE,
        genom_state = genom_state
    )
}

annot_ers <- add_expressed_genes(tissue = "lung", gtf_path = gtf_path, annot_ers = annot_ers1)
refined_ers <- refine_ERs(annot_ers)

refined_ers
}
