% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_options}
\alias{tissue_options}
\title{The different tissues that can be filtered on for gene expression}
\format{
A character vector with all of the tissue options available to filter
on. These are to be used in conjunction with the
\code{\link{add_expressed_genes}} function.
}
\source{
See tissues.R in data-raw
}
\usage{
tissue_options
}
\description{
These options were derived from the contents of the GTEx analysis gene median
RPKM file.
}
\keyword{datasets}
