% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressed_genes.R
\name{add_expressed_genes}
\alias{add_expressed_genes}
\title{Adding the nearest expressed genes}
\usage{
add_expressed_genes(
  input_file = NULL,
  tissue,
  gtf_path,
  species = "Homo_sapiens",
  annot_ers
)
}
\value{
Granges with annotated ERs and details of their nearest expressed
genes
}
\description{
Updating expressed regions with the expressed gene that is closest to it.
After entering the tissue that has been sequenced, the nearest gene and
nearest expressed gene will be added to the metadata columns of the annotated
ERs.
}
\examples{

if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    gtf_path <- .file_cache(gtf_url)
}

ex_opt_ers <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle(c("chr21", "chr22"), c(2, 2)),
    ranges = IRanges::IRanges(
        start = c(5116369, 5118691, 5125879, 5128214),
        end = c(5117231, 5118847, 5125988, 5128403)
    )
)

ex_opt_ers_w_exp_genes <- add_expressed_genes(
    tissue = "lung", gtf_path = gtf_path,
    annot_ers = ex_opt_ers
)

ex_opt_ers_w_exp_genes
}
