% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ers_delta_example}
\alias{ers_delta_example}
\title{An example set of ER deltas}
\format{
A tibble/dataframe with the sums, means, medians, n_eq_0 and
propor_eq_0 for each combination of mccs (5 & 10) and mrgs (10 & 20)
}
\source{
See example.R in data-raw
}
\usage{
ers_delta_example
}
\description{
This set of deltas was calculated using ers_example and exons from ensembl.
}
\keyword{datasets}
