% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coverage.R
\name{get_chr_info}
\alias{get_chr_info}
\title{Get information from UCSC about the chromosomes passed in}
\usage{
get_chr_info(chrs, genome)
}
\arguments{
\item{chrs}{chromosomes to look up (must match UCSC format)}

\item{genome}{the UCSC genome to look at}
}
\value{
a dataframe with data on the passed in chromosomes
}
\description{
Get information from UCSC about the chromosomes passed in
}
\examples{
eg_info <- get_chr_info(chrs = c("chr21", "chr22"), genome = "hg38")
}
