% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_count_matrix.R
\name{get_count_matrix}
\alias{get_count_matrix}
\title{Generate the count matrix}
\usage{
get_count_matrix(bw_paths, annot_ers, cols = NULL)
}
\arguments{
\item{bw_paths}{Vector containing the bigwig file paths to read in}

\item{annot_ers}{GRangesList containing the annotated ERs (product of annotatERs())}

\item{cols}{A dataframe containing the information to be used as colData for the
output. If NULL then the bw_paths will be used for the colData}
}
\value{
A Ranged Summarized Experiment containing the gene counts as an assay
}
\description{
Scores the mean coverage of the expressed regions as a count matrix
}
\examples{
megadepth::install_megadepth()
if (!exists("rec_url")) {
    rec_url <- recount::download_study(
        project = "SRP012682",
        type = "samples",
        download = FALSE
    ) # .file_cache is an internal function to download a bigwig file from a link
    # if the file has been downloaded recently, it will be retrieved from a cache
}
bw_path <- .file_cache(rec_url[1])

ex_opt_ers <- GenomicRanges::GRanges( # this is created to not overlap
    seqnames = S4Vectors::Rle(c("chr1", "chr2"), c(4, 1)),
    ranges = IRanges::IRanges(
        start = c(1:5),
        end = seq(100, 500, 100)
    )
)

example_cm <- get_count_matrix(bw_paths = c(bw_path, bw_path), annot_ers = ex_opt_ers)
example_cm
}
