% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ers.R
\name{get_ers}
\alias{get_ers}
\title{Define sets of ERs}
\usage{
get_ers(coverage, mccs, mrgs)
}
\arguments{
\item{coverage}{the coverage of the bigwig files passed into \code{\link{get_coverage}}.}

\item{mccs}{mean coverage cut-offs to apply.}

\item{mrgs}{max region gaps to apply.}
}
\value{
list containing sets of ERs, each generated using a particular
combination of MCC and MRG.
}
\description{
\code{gen_ERs} defines ERs across an inputted range of mean coverage cut-offs
(MCCs) and max region gaps (MRGs).
}
\examples{

eg_ers <- get_ers(coverage = coverage_example, mccs = c(5, 10), mrgs = c(10, 20))
}
