% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-ERs.R
\name{get_ers_delta}
\alias{get_ers_delta}
\title{Calculates delta for sets of ERs}
\usage{
get_ers_delta(ers, opt_exons, delta_fun = ODER:::.delta)
}
\arguments{
\item{ers}{Sets of ERs across various MCCs/MRGs - output of
\code{\link{get_ers}}.}

\item{opt_exons}{GRanges object that contains the regions that ideally, you
want the ER definitions to match - output of \code{\link{get_exons}}.}

\item{delta_fun}{Function that calculates the delta between ERs and
\code{opt_exons}. Takes as input a set of ERs from \code{ers} and
\code{opt_exons}. Then outputs a tibble/dataframe containing the summarised
delta scores for that set of one set of ERs.}
}
\value{
tibble/dataframe containing summarised delta values. One row per set
of ERs.
}
\description{
Calculates the median exon delta and the number of ERs with an exon delta of
0 by comparing each combination of MCC and MRG with the optimum exons from
the ensembl database.
}
\examples{
\dontshow{
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
gtf_path <- file_cache(gtf_url)
# file_cache is an internal function to download a bigwig file from a
# link
# if the file has been downloaded recently, it will be retrieved from a
# cache gtf_path <- file_cache(gtf_url)


if (!exists("eg_opt_exons")) {
    eg_opt_exons <- get_exons(
        gtf = gtf_path,
        ucsc_chr = TRUE,
        ignore.strand = TRUE
    )
}
}
if (!exists("eg_ers_delta")) {
    eg_ers_delta <- get_ers_delta(
        ers = ers_example, # ers_example is from the package data folder
        opt_exons = eg_opt_exons,
        delta_fun = .delta
    ) # .delta is ODER's default, you can pass in your own if you have one
}
message(eg_ers_delta)
}
