% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise-ERs.R
\name{get_exons}
\alias{get_exons}
\title{Obtain set of non-overlapping exons}
\usage{
get_exons(gtf, ucsc_chr, ignore.strand = TRUE, biotype = "Non-overlapping")
}
\arguments{
\item{gtf}{Either a string containg the path to a .gtf file or a pre-imported
gtf using \code{rtracklayer::import} .}

\item{ucsc_chr}{logical scalar, determining whether to add "chr" prefix to
the seqnames of non-overlapping exons and change "chrMT" -> "chrM". Note,
if set to TRUE and seqnames already have "chr", it will not add another.}

\item{ignore.strand}{logical value for input into
\code{\link[GenomicRanges]{findOverlaps}}, default is True.}

\item{biotype}{Filters the GTF file passed in to what would be considered the
"Gold Standard" exons. The Default is "Non-overlapping" but the options are:
"Non-overlapping" (exons that don't intersect each other),
"Three Prime" (3' UTR), "Five Prime" (5' UTR), "Internal" (Internal coding),
"lncRNA" (Long Non-Coding RNA), "ncRNA" (Non-Coding RNA) and "Pseudogene"}
}
\value{
GRanges object containing non-overlapping exons.
}
\description{
Downloads a well-defined set of exons to be used in obtaining the optimum set
of Expressed regions. These exons are used in calculating the exon deltas.
}
\examples{
\dontshow{
if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    gtf_path <- file_cache(gtf_url)
}

if (!exists("gtf_gr")) {
    gtf_gr <- rtracklayer::import(gtf_path)
}
}
if (!exists("eg_opt_exons")) {
    eg_opt_exons <- get_exons(
        gtf = gtf_gr,
        ucsc_chr = TRUE,
        ignore.strand = TRUE
    )
}

eg_opt_exons
}
