% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatER.R
\name{get_junctions}
\alias{get_junctions}
\title{Get junctions that overlap the optimally defined ERs}
\usage{
get_junctions(opt_ers, junc_data, gtf_path)
}
\arguments{
\item{opt_ers}{optimally defined ERs (the product of the ODER function)}

\item{junc_data}{junction data that should match the ERs passed into opt_ers}

\item{gtf_path}{a gtf file or txdb with exon data}
}
\value{
optimally defined ers annotated with junction and gene information
}
\description{
The junctions will be added to the ERs passed in as a metadata column, a
\code{GRangeslist} with each GRanges element of the list corresponding to it's
associated ER. If there are no overlaps, the \code{GRangeslist} will be empty. Any
ERs that do not have an overlapping junction will have an empty GRanges. The
respective genes that each ER could be associated with will also be passed in
as a metadata column, a character list.
}
\examples{
\dontshow{
if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    # .file_cache is an internal function to download a bigwig file from a link
    # if the file has been downloaded recently, it will be retrieved from a cache
    gtf_path <- .file_cache(gtf_url)
}

if (!exists("gtf_gr")) {
    gtf_gr <- rtracklayer::import(gtf_path)
}

chrs_to_keep <- c("21", "22")
#' if (!exists("ens_txdb")) {
#### preparing the txdb object
hg38_chrominfo <- GenomeInfoDb::getChromInfoFromUCSC("hg38")
new_info <- hg38_chrominfo$size[match(
    chrs_to_keep,
    GenomeInfoDb::mapSeqlevels(hg38_chrominfo$chrom, "Ensembl")
)]
names(new_info) <- chrs_to_keep
gtf_gr_tx <- GenomeInfoDb::keepSeqlevels(gtf_gr,
    chrs_to_keep,
    pruning.mode = "tidy"
)
GenomeInfoDb::seqlengths(gtf_gr_tx) <- new_info
GenomeInfoDb::seqlevelsStyle(gtf_gr_tx) <- "UCSC"
rtracklayer::genome(gtf_gr_tx) <- "hg38"

ens_txdb <- GenomicFeatures::makeTxDbFromGRanges(gtf_gr_tx)
GenomeInfoDb::seqlevelsStyle(ens_txdb) <- "Ensembl"
################### end of txdb creation

}
}
example_ers <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle(c("chr21", "chr22"), c(10, 1)),
    ranges = IRanges::IRanges(
        start = c(
            5026423, 24738, 5032218, 5033895, 17554, 50800446,
            50800539, 16570, 50800790, 15005, 20312
        ),
        end = c(
            5323718, 24891, 5033407, 5033980, 17728, 50800910,
            50800817, 16723, 50800910, 15038, 20582
        ),
        names = head(letters, 11)
    ),
    strand = S4Vectors::Rle((c("+", "-")), c(6, 5)),
    score = 1:11,
    GC = seq(1, 0, length = 11)
)

example_junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)

example_er_juncs <- get_junctions(
    opt_ers = example_ers,
    junc_data = example_junctions,
    gtf_path = ens_txdb #can either be a gtf file or txdb in the Ensembl format
)

print(example_er_juncs)
}
\seealso{
dasper::junction_annot
}
