% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatER.R
\name{get_junctions}
\alias{get_junctions}
\title{Get junctions that overlap the optimally defined ERs}
\usage{
get_junctions(opt_ers, junc_data, gtf_path)
}
\arguments{
\item{opt_ers}{optimally defined ERs (the product of the ODER function)}

\item{junc_data}{junction data that should match the ERs passed into opt_ers}

\item{gtf_path}{a gtf file with exon data}
}
\value{
optimally defined ers annotated with junction and gene information
}
\description{
The junctions will be added to the ERs passed in as a metadata column, a
GRangeslist with each GRanges element of the list corresponding to it's
associated ER. If there are no overlaps, the GRangeslist will be empty.
Any ERs that do not have an overlapping junction will have an empty GRanges.
The respective genes that each ER could be associated with will also be passed
in as a metadata column, a character list
}
\details{
Each Granges of the GRangeslist will have the metadata columns of "in_ref",
"gene_id_start", "tx_name_start", "exon_name_start", "strand_start",
"exon_width_start","gene_id_end", "tx_name_end", "exon_name_end", "strand_end",
"exon_width_end", "gene_id_junction", "strand_junction", "type" and "er_index
added by dasper's junction_annot.
}
\examples{
\dontshow{
if (!exists("gtf_path")) {
    gtf_url <- paste0(
        "http://ftp.ensembl.org/pub/release-103/gtf/",
        "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
    )
    # .file_cache is an internal function to download a bigwig file from a link
    # if the file has been downloaded recently, it will be retrieved from a cache
    gtf_path <- .file_cache(gtf_url)
}
}
example_ers <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle(c("chr21", "chr22"), c(10, 1)),
    ranges = IRanges::IRanges(
        start = c(
            5026423, 24738, 5032218, 5033895, 17554, 50800446,
            50800539, 16570, 50800790, 15005, 20312
        ),
        end = c(
            5323718, 24891, 5033407, 5033980, 17728, 50800910,
            50800817, 16723, 50800910, 15038, 20582
        ),
        names = head(letters, 11)
    ),
    strand = S4Vectors::Rle((c("+", "-")), c(6, 5)),
    score = 1:11,
    GC = seq(1, 0, length = 11)
)

example_junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)

example_er_juncs <- get_junctions(
    opt_ers = example_ers,
    junc_data = example_junctions,
    gtf_path = gtf_path
)

print(example_er_juncs)
}
\seealso{
dasper::junction_annot
}
