% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_ERs.R
\name{get_no_overlap_exons}
\alias{get_no_overlap_exons}
\title{Optain set of non-overlapping exons}
\usage{
get_no_overlap_exons(gtf, ucsc_chr, ignore.strand)
}
\arguments{
\item{gtf}{Either a string if path to a .gtf file or a pre-imported gtf using
\code{\link[rtracklayer]{import}}.}

\item{ucsc_chr}{logical scalar, determining whether to add "chr" prefix to the
seqnames of non-overlapping exons and change "chrMT" -> "chrM". Note, if
set to TRUE and seqnames already have "chr", it will not add another.}

\item{ignore.strand}{logical value for input into
\code{\link[GenomicRanges]{findOverlaps}}.}
}
\value{
GRanges object containing non-overlapping exons.
}
\description{
\code{get_no_overlap_exons} uses as input a gtf to optain set of
non-overlapping exons. These can be used in \code{\link{gen_ERs}} as the
\code{opt_gr} to optimise ER definitions.
}
