% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_ERs.R
\name{get_opt_ERs}
\alias{get_opt_ERs}
\title{Optains optimised set of ERs}
\usage{
get_opt_ERs(ERs_MCCs_MRGs, delta_df)
}
\arguments{
\item{ERs_MCCs_MRGs}{Sets of ERs across various MCCs/MRGs - output of
\code{\link{gen_ERs}}.}

\item{delta_df}{Output of \code{\link{calc_ERs_delta}}.}
}
\value{
list containing optimised ERs, optimal pair of MCC/MRGs and
  \code{delta_df}
}
\description{
\code{get_opt_ERs} optains
}
