% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODER.R
\name{ODER}
\alias{ODER}
\title{Optimising the Definition of Expressed Regions}
\usage{
ODER(
  bw_paths,
  aucs,
  target_auc,
  chr_to_filter = stringr::str_c("chr", c(1:22, "X", "Y", "M")),
  MCCs,
  MRGs,
  gtf = NULL,
  ucsc_chr,
  ignore.strand,
  opt_gr = NULL
)
}
\arguments{
\item{bw_paths}{paths to bigwig files.}

\item{aucs}{vector containing aucs matching the order of bigwig paths.}

\item{target_auc}{total auc to normalise all samples to. E.g. 40e6 * 100
would be the estimated total auc for sample sequenced to 40 million reads
of 100bp in length.}

\item{chr_to_filter}{chrs to obtain mean coverage for.}

\item{MCCs}{mean coverage cut-offs to apply.}

\item{MRGs}{max region gaps to apply.}

\item{gtf}{Either a string if path to a .gtf file or a pre-imported gtf using
\code{\link[rtracklayer]{import}}.}

\item{ucsc_chr}{logical scalar, determining whether to add "chr" prefix to the
seqnames of non-overlapping exons and change "chrMT" -> "chrM". Note, if
set to TRUE and seqnames already have "chr", it will not add another.}

\item{ignore.strand}{logical value for input into
\code{\link[GenomicRanges]{findOverlaps}}.}

\item{opt_gr}{GRanges object that contains the regions that ideally, you want
to the ER definitions to match}
}
\value{
list containing optimised ERs, optimal pair of MCC/MRGs and
  \code{delta_df}
}
\description{
\code{ODER} is a wrapper function that runs the pipeline from generation of ERs
across different MCCs/MRGs to selecting the optimal set of ERs. Either a gtf
is required, to optain a set of non-overlapping exons to use as the
\code{opt_gr} OR users can provide their own set of ranges to optimise ER
definitions towards.
}
