% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ERs.R
\name{plot_ers}
\alias{plot_ers}
\title{Plot Expressed regions}
\usage{
plot_ers(ers_delta, opt_mcc_mrg)
}
\arguments{
\item{ers_delta}{tibble/dataframe containing summarised delta values. One row per set
of ERs.}

\item{opt_mcc_mrg}{vector containing the optimum mcc and mrg, in that order}
}
\value{
Plot of MCC against median delta and number of ERS with a delta of 0
}
\description{
Plot Expressed regions
}
\examples{
# ers_delta_example is from the package data folder
eg_plots <- plot_ers(ers_delta = ers_delta_example, opt_mcc_mrg = c("mcc_10", "mrg_20"))

eg_plots
}
