% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine-ERs.R
\name{refine_ERs}
\alias{refine_ERs}
\title{Refines the ERs start and end points}
\usage{
refine_ERs(annot_ers)
}
\arguments{
\item{annot_ers}{ERs that have been annotated (result of annotatER)}
}
\value{
Genomic ranges with refined base pair starts and ends and a logical
vector listing changes
}
\description{
Uses the junctions in the metadata to modify the starts and ends of the
Expressed regions
}
\examples{
\dontshow{
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
) # .file_cache is an internal function to download a bigwig file from a link
# if the file has been downloaded recently, it will be retrieved from a cache

bw_path <- ODER:::.file_cache(url[1])
gtf_url <- paste0(
    "http://ftp.ensembl.org/pub/release-103/gtf/",
    "homo_sapiens/Homo_sapiens.GRCh38.103.chr.gtf.gz"
)
gtf_path <- ODER:::.file_cache(gtf_url)
}

opt_ers <- ODER(
    bw_paths = bw_path, auc_raw = auc_example,
    auc_target = 40e6 * 100, chrs = c("chr21", "chr22"),
    genome = "hg38", mccs = c(5, 10), mrgs = c(10, 20),
    gtf = gtf_path, ucsc_chr = TRUE, ignore.strand = TRUE,
    exons_no_overlap = NULL, bw_chr = "chr"
)

junctions <- SummarizedExperiment::rowRanges(dasper::junctions_example)
annot_ers <- annotatERs(
    opt_ers = opt_ers[["opt_ers"]], junc_data = junctions,
    gtf_path = gtf_path, chrs_to_keep = c("21", "22"), ensembl = TRUE
)

refined_ers <- refine_ERs(annot_ers)

refined_ers
}
